refactor(hooks): 특정 사용처에서만 사용되는 공용 hooks 디렉토리 hook을 사용처로 이동
